refactor: rename file_mod occurences to filemod
