+++ src/diffhouse/entities/__init__.py
@@ -3,7 +3,7 @@
 from diffhouse.entities.branch import Branch
 from diffhouse.entities.commit import Commit
 from diffhouse.entities.diff import Diff
-from diffhouse.entities.file_mod import FileMod
+from diffhouse.entities.filemod import FileMod
 from diffhouse.entities.tag import Tag
 
 __all__ = ['Commit', 'FileMod', 'Diff', 'Branch', 'Tag']
